quartz-dashboard/flask-app/requirements.txt
mysql-connector-python
flask-mongoengine
flask-cors
sslscan
gitpython
urllib3==1.26.18
xmltodict
requests
docker
pexpect
pytest